+++ src/components/NavMenuItem.vue
@@ -1,6 +1,9 @@
 <template>
   <div class="nav-item">
-    {{ title }}
+    <div class="icon">
+      <slot></slot>
+    </div>
+    <div v-if="!$store.getters.getNavigationSize" class="title">{{ title }}</div>
   </div>
 </template>
 
@@ -11,21 +14,39 @@ export default {
     title: {
       default: 'No Title',
       type: String
+    },
+    icon: {
+      default: 'I',
+      type: String
     }
   }
 }
 </script>
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
+<style lang="scss" scoped>
 .nav-item {
   width: 100%;
-  height: 5.5rem;
-  line-height: 6rem;
+  height: 3.5rem;
+  line-height: 3.5rem;
   text-align: center;
   border-bottom: 1px solid #373144;
   cursor: pointer;
+  margin: 0;
+
+  > div {
+    display: inline-block;
+    vertical-align: middle;
+  }
+
+  > .icon {font-size: 1rem; width: 1rem; height: 2rem; margin-right: 0.75rem; margin-left: 0.5rem;
+    > i {
+      vertical-align: top;
+      margin-top: 0.4rem;
+    }
+  }
+  > .text {}
 }
 
-.nav-item:hover {background-color: rgba(255,255,255,0.05);}
+.nav-item:hover {background-color: rgba(0,0,0,0.05);}
 </style>