+++ src/views/authenticated/Navigation.vue
@@ -1,8 +1,11 @@
 <template>
-    <div class="navigation">
-      <NavMenuItem @click="navigateTo('dashboard')" link="users" title="Dashboard"/>
-      <NavMenuItem @click="navigateTo('user')" link="users" title="User"/>
-      <NavMenuItem @click="logout_user" title="Logout"/>
+    <div class="navigation p-shadow-4">
+      <NavMenuItem @click="navigateTo('dashboard')" link="users" title="Dashboard" icon="I">
+        <i class="pi pi-check"></i>
+      </NavMenuItem>
+      <NavMenuItem @click="navigateTo('user')" link="users" title="User" icon="I"/>
+      <NavMenuItem @click="logout_user" title="Logout" icon="I"/>
+      <NavMenuItem @click="toggleWidth" title="W" icon="I"/>
   </div>
 </template>
 
@@ -17,7 +20,9 @@ export default {
     msg: String
   },
   methods: {
-
+    toggleWidth: function() {
+      this.$store.dispatch('TOGGLE_NAVIGATION_SIZE');
+    }
   }
 }
 </script>
@@ -25,9 +30,11 @@ export default {
 <style lang="scss" scoped>
 @import '@/assets/theme/main.scss';
 
-.navigation {width: 100%;
-            text-align: left;
-            height: 100vh;
-            box-shadow: 0px 2px 20px 2px rgba(0, 0, 0, 0.1);
-            background-color: $main_purple;}
+.navigation {
+  width: 100%;
+  color: #333333;
+  text-align: left;
+  height: 100vh;
+  background: #ffffff;
+  }
 </style>