+++ src/views/authenticated/Root.vue
@@ -1,21 +1,24 @@
 <template>
   <div class="auth-root">
-    <Navigation class="navigation-pane"/>
-    <div class="content-pane">
+    <Navigation :class="['navigation-pane', compact_navigation_class]"/>
+    <div :class="['content-pane', compact_navigation_class]">
       <router-view></router-view>
     </div>
   </div>
 </template>
 
 <script>
-import { RouterView } from 'vue-router';
 import Navigation from '@/views/authenticated/Navigation.vue';
 
 export default {
   name: 'AuthRoot',
   components: {
     Navigation,
-    'router-view' : RouterView
+  },
+  computed: {
+    compact_navigation_class: function(){
+      return (this.$store.getters.getNavigationSize) ? 'compact' : '';
+    }
   },
   props: {
     msg: String
@@ -26,30 +29,45 @@ export default {
 <style lang="scss" scoped>
 @import '@/assets/theme/main.scss';
 
-$navigation_width: 7rem;
+$navigation_width: 10rem;
+$navigation_width_compact: 4rem;
 
 .auth-root {
   width: 100vw;
   min-height: 100vh;
   text-align: left;
   font-size: 0;
-  background-color: yellow;
+  background-color: rgb(238, 238, 238);
+  position: relative;
 
   .content-pane {
     width: calc(100vw - #{$navigation_width});
     vertical-align: top;
     text-align: left;
     display: inline-block;
-    background-color: rgb(238, 238, 238);
+    //background-color: rgb(238, 238, 238);
     min-height: 100vh;
     font-size: 1rem;
+    z-index: 9;
+  }
+
+  .content-pane.compact {
+    width: calc(100vw - #{$navigation_width_compact});
   }
 
   .navigation-pane {
     vertical-align: top;
     font-size: 1rem;
     display: inline-block;
+    width: $navigation_width;
     max-width: $navigation_width;
+    box-shadow: 5px 5px 20px 5px rgba(0,0,0, 0.05);
+    z-index: 10;
+  }
+
+  .navigation-pane.compact {
+    width: $navigation_width_compact;
+    max-width: $navigation_width_compact;
   }
 }