+++ src/views/public/Navigation.vue
@@ -1,7 +1,12 @@
 <template>
   <div class="navigation">
 
-    <button @click="login" class="login">Login</button>
+    <div class="nav-buttons-group">
+      <Button class="p-button-text p-m-2 p-mt-3 nav-button" label="Reset" @click="navigateTo('password_reset')" />
+      <Button class="p-button-text p-m-2 p-mt-3 nav-button" label="Home" @click="navigateTo('home')" />
+      <Button class="p-button-text p-m-2 p-mt-3 nav-button" label="Register" @click="navigateTo('register')" />
+      <Button class="p-button-text p-m-2 p-mt-3 nav-button" label="Login" @click="navigateTo('login')" />
+    </div>
 
   </div>
 </template>
@@ -15,9 +20,7 @@ export default {
     msg: String
   },
   methods: {
-    login: function() {
-      this.navigateTo('login');
-    }
+
   }
 }
 </script>
@@ -29,15 +32,17 @@ export default {
   width: 100vw;
   text-align: left;
   color: #ffffff;
-  height: 4rem;
+  height: 3.75rem;
   box-shadow: 0px 2px 20px 2px rgba(0, 0, 0, 0.1);
-  background-color: #2a2338;
 
-  .login {
+  .nav-buttons-group {
     float: right;
-    margin-top: 0.5rem;
-    margin-right: 1rem;
-    height: 3rem;
+
+    .nav-button {
+      margin-top: 0.5rem;
+      margin-right: 1rem;
+      font-size: 0.85rem;
+    }
   }
 }
 </style>