+++ src/views/public/Root.vue
@@ -2,17 +2,20 @@
   <div class="public-root">
     <Navigation />
     <RouterView></RouterView>
+
+    <HomePage v-if="$route.name === 'home'"/>
   </div>
 </template>
 
 <script>
-import { RouterView } from 'vue-router'
 import Login from '@/components/Login.vue'
 import Navigation from '@/views/public/Navigation.vue';
+import HomePage from "@/views/public/Home.vue";
 
 export default {
   name: 'UnauthenticatedRoot',
   components: {
+    HomePage,
     Login,
     Navigation
   },
@@ -31,6 +34,5 @@ export default {
             padding: 5rem;
             text-align: left;
             max-width: 30rem;
-            height: 100vh;
-            background-color: #221d2d;}
+            height: 100vh;}
 </style>