+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### Added
 - Base Application
 - index.html, App.vue, main.js
+- Login Page
+- Password Reset
+- Account Recovery
+- Account Registration
 ### Changed
 ### Fixed
 ### Removed